pyproject.toml
[tool.poetry]
name = "user-pets"
version = "0.1.0"
description = ""
authors = ["Vas <vas@va.va>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-cors-headers = "^4.1.0"
djangorestframework-simplejwt = "^5.2.2"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.6"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
pillow = "^10.0.0"
django-stubs = "^4.2.3"
django-filter = "^23.2"
coreapi = "^2.3.3"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"
